+++ src/pdo/sink.rs
@@ -165,7 +165,7 @@ bitfield! {
     /// Minimum voltage in 50 mV units
     pub u16, min_voltage, set_min_voltage: 19, 10;
     /// Operational power in 250 mW units
-    pub u16, operational_power, set_operational_power: 9, 0;
+    pub u32, operational_power, set_operational_power: 9, 0;
 }
 
 /// Battery supply data
@@ -177,7 +177,7 @@ pub struct BatteryData {
     /// Minimum voltage in mV
     pub min_voltage_mv: u16,
     /// Operational power in mW
-    pub operational_power_mw: u16,
+    pub operational_power_mw: u32,
 }
 
 impl TryFrom<u32> for BatteryData {
@@ -331,7 +331,7 @@ bitfield! {
     /// Minimum voltage in 100mV units
     pub u16, min_voltage, set_min_voltage: 15, 8;
     /// PDP in 1W units
-    pub u16, pdp, set_pdp: 7, 0;
+    pub u32, pdp, set_pdp: 7, 0;
 }
 
 /// EPR Adjustable voltage supply data
@@ -343,7 +343,7 @@ pub struct EprAvsData {
     /// Minimum voltage in mV
     pub min_voltage_mv: u16,
     /// PDP in mW
-    pub pdp_mw: u16,
+    pub pdp_mw: u32,
 }
 
 impl TryFrom<u32> for EprAvsData {
